Split code based on routes and fix sample requests
