+++ .gitignore
@@ -1,2 +1,4 @@
 .python-version
+.vscode
 venv/
+__pycache__/
