+++ routes/OHLC.py
@@ -1,0 +1,51 @@
+from flask import Blueprint
+from flask_cors import CORS, cross_origin
+from flask import request
+import yfinance as yf
+from flask import jsonify
+
+
+OHLC = Blueprint('OHLC', __name__)
+
+### Sample request body for /
+
+# {
+#     "context": {
+#         "ticker": "MSFT",
+#         "start": "2018-01-01",
+#         "end": "2020-11-30"
+#     }
+# }
+
+
+@OHLC.route("/", methods=["POST"])
+@cross_origin()
+def solve():
+    context = request.json["context"]
+    if request.method == "POST":
+
+        values = yf.Ticker(context["ticker"]).history(
+            start=context["start"], end=context["end"]
+        )
+
+        res = {}
+
+        for i in range(0, values.shape[0]):
+
+            res[i] = {
+                "date": values.iloc[i].name.strftime("%Y-%m-%d"),
+                "open": values.iloc[i]["Open"],
+                "high": values.iloc[i]["High"],
+                "low": values.iloc[i]["Low"],
+                "close": values.iloc[i]["Close"],
+                "volume": values.iloc[i]["Volume"],
+                "dividends": values.iloc[i]["Dividends"],
+                "stock splits": values.iloc[i]["Stock Splits"],
+            }
+
+        data = {
+            "length": values.shape[0],
+            "data": res,
+        }
+
+        return jsonify(data)
